requirements.txt
streamlit>=1.32.0
google-cloud-aiplatform>=1.47.0
google-cloud-bigquery>=3.17.0
vertexai>=1.47.0
plotly>=5.19.0
pandas>=2.0.0
python-dotenv>=1.0.0
db-dtypes>=1.2.0
